Kafka addition
