+++ metricbeat/module/kafka/broker/broker_integration_test.go
@@ -2,22 +2,15 @@
 package broker
 
 import (
-	"fmt"
-	"os"
 	"testing"
 
-	"github.com/Shopify/sarama"
 	mbtest "github.com/elastic/beats/metricbeat/mb/testing"
-)
-
-const (
-	kafkaDefaultHost = "localhost"
-	kafkaDefaultPort = "9092"
+	"github.com/elastic/beats/metricbeat/module/kafka"
 )
 
 func TestData(t *testing.T) {
 
-	generateKafkaData(t)
+	kafka.GenerateKafkaData(t)
 
 	f := mbtest.NewEventsFetcher(t, getConfig())
 	err := mbtest.WriteEvents(f, t)
@@ -30,74 +23,34 @@ func TestData(t *testing.T) {
 func TestTopic(t *testing.T) {
 
 	// Create initial topic
-	generateKafkaData(t)
+	kafka.GenerateKafkaData(t)
 
-	/*f := mbtest.NewEventsFetcher(t, getConfig())
-	dataBefore, err := f.Fetch()
+	f := mbtest.NewEventsFetcher(t, getConfig())
+	data, err := f.Fetch()
 	if err != nil {
 		t.Fatal("write", err)
 	}
 
-	// Create 10 messages
-	for i := 0; i < 10; i++ {
-		generateKafkaData(t)
-	}
+	t.Errorf("DATA: %+v", data)
 
-	dataAfter, err := f.Fetch()
-	if err != nil {
-		t.Fatal("write", err)
-	}*/
+	/*
 
-}
+		// Create 10 messages
+		for i := 0; i < 10; i++ {
+			generateKafkaData(t)
+		}
 
-func generateKafkaData(t *testing.T) {
-
-	config := sarama.NewConfig()
-	client, err := sarama.NewClient([]string{getTestKafkaHost()}, config)
-	if err != nil {
-		t.Errorf("%s", err)
-	}
-
-	producer, err := sarama.NewSyncProducerFromClient(client)
-	if err != nil {
-		t.Error(err)
-	}
-	defer producer.Close()
-
-	msg := &sarama.ProducerMessage{
-		Topic: "testtopic",
-		Value: sarama.StringEncoder("Hello World"),
-	}
-
-	_, _, err = producer.SendMessage(msg)
-	if err != nil {
-		t.Errorf("FAILED to send message: %s\n", err)
-	}
+		dataAfter, err := f.Fetch()
+		if err != nil {
+			t.Fatal("write", err)
+		}*/
 
 }
 
 func getConfig() map[string]interface{} {
 	return map[string]interface{}{
 		"module":     "kafka",
-		"metricsets": []string{"topic"},
-		"hosts":      []string{getTestKafkaHost()},
+		"metricsets": []string{"broker"},
+		"hosts":      []string{kafka.GetTestKafkaHost()},
 	}
-}
-
-func getTestKafkaHost() string {
-	return fmt.Sprintf("%v:%v",
-		getenv("KAFKA_HOST", kafkaDefaultHost),
-		getenv("KAFKA_PORT", kafkaDefaultPort),
-	)
-}
-
-func getenv(name, defaultValue string) string {
-	return strDefault(os.Getenv(name), defaultValue)
-}
-
-func strDefault(a, defaults string) string {
-	if len(a) == 0 {
-		return defaults
-	}
-	return a
 }
